requirements.txt
capstone
pyelftools
pyyaml
rich
macholib

